+++ src/main/scala/smptk/io/ImageIO.scala
@@ -53,7 +53,7 @@ object ImageIO {
             } else if (imageData.pixelDimensionality != 1) {
               Failure(new Exception("wrong pixel dimensionality in image data"))
             } else {
-              val domain = DiscreteImageDomain1D(imageData.origin(0).toFloat, imageData.spacing(0).toFloat,imageData.size(0).toInt)  
+              val domain = DiscreteImageDomain1D(imageData.origin(0), imageData.spacing(0),imageData.size(0).toInt)  
               Success(DiscreteScalarImage1D(domain, imageData.data))
             }
           }
@@ -77,8 +77,8 @@ object ImageIO {
               Failure(new Exception("wrong pixel dimensionality in image data"))
             } else {
               val domain = DiscreteImageDomain3D(
-                (imageData.origin(0).toFloat, imageData.origin(1).toFloat, imageData.origin(2).toFloat),
-                (imageData.spacing(0).toFloat, imageData.spacing(1).toFloat, imageData.spacing(2).toFloat),
+                (imageData.origin(0), imageData.origin(1), imageData.origin(2)),
+                (imageData.spacing(0), imageData.spacing(1), imageData.spacing(2)),
                 (imageData.size(0).toInt, imageData.size(1).toInt, imageData.size(2).toInt))
               Success(DiscreteScalarImage3D(domain, imageData.data))
             }
@@ -103,8 +103,8 @@ object ImageIO {
               Failure(new Exception("wrong pixel dimensionality in image data"))
             } else {
               val domain = DiscreteImageDomain2D(
-                (imageData.origin(0).toFloat, imageData.origin(1).toFloat),
-                (imageData.spacing(0).toFloat, imageData.spacing(1).toFloat),
+                (imageData.origin(0), imageData.origin(1)),
+                (imageData.spacing(0), imageData.spacing(1)),
                 (imageData.size(0).toInt, imageData.size(1).toInt))
               Success(DiscreteScalarImage2D(domain, imageData.data))
             }
